pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muviz"
version = "0.1.0"
description = "Audio visualization video generator"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Muviz Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.0.0",
    "librosa>=0.10.0",
    "numpy>=1.21.0",
    "moviepy>=1.0.3",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "soundfile>=0.12.0",
]

[project.scripts]
muviz = "muviz.cli:main"

[project.urls]
Homepage = "https://github.com/muviz/muviz"
